% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{listvector}
\alias{listvector}
\title{Create vector of mode list of length n.}
\usage{
listvector(n = 1)
}
\arguments{
\item{n}{Length of vector.}
}
\value{
Vector of mode list of length n.
}
\description{
Create vector of mode list of length n.
}
\examples{

## Simulate data and get initial model parameter estimates
A <- listvector (n=3)

}
\keyword{list}
\keyword{vector}
